creacion de CartaServices y de CartaController para probar si funcionan los endpoints

+++ PokeTrade/src/main/java/poketrade/PokeTrade/repository/CartaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartaRepository extends JpaRepository<Carta, Long> {
+public interface CartaRepository extends JpaRepository<Carta, Integer> {
 }
